version: move PredecessorVersion out of clusterupgrade package
Release note: None
