+++ pkg/cmd/roachtest/roachtestutil/clusterupgrade/BUILD.bazel
@@ -3,11 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "clusterupgrade",
-    srcs = [
-        "clusterupgrade.go",
-        "predecessor_version.go",
-    ],
-    embedsrcs = ["predecessor_version.json"],
+    srcs = ["clusterupgrade.go"],
     importpath = "github.com/cockroachdb/cockroach/pkg/cmd/roachtest/roachtestutil/clusterupgrade",
     visibility = ["//visibility:public"],
     deps = [
@@ -19,7 +15,6 @@ go_library(
         "//pkg/roachprod/logger",
         "//pkg/util/retry",
         "//pkg/util/version",
-        "@com_github_cockroachdb_errors//:errors",
     ],
 )
