+++ pkg/cmd/roachtest/roachtestutil/mixedversion/planner_test.go
@@ -51,7 +51,7 @@ var (
 		return *bv
 	}()
 	previousVersion = func() string {
-		pv, err := clusterupgrade.PredecessorVersion(buildVersion)
+		pv, err := version.PredecessorVersion(buildVersion)
 		if err != nil {
 			panic(err)
 		}
