+++ pkg/cmd/roachtest/tests/multitenant_upgrade.go
@@ -69,7 +69,7 @@ func registerMultiTenantUpgrade(r registry.Registry) {
 //   - Tenant14{Binary: Cur, Cluster: Cur}: Create tenant 14 and verify it works.
 //   - Tenant12{Binary: Cur, Cluster: Cur}: Restart tenant 14 and make sure it still works.
 func runMultiTenantUpgrade(ctx context.Context, t test.Test, c cluster.Cluster, v version.Version) {
-	predecessor, err := clusterupgrade.PredecessorVersion(v)
+	predecessor, err := version.PredecessorVersion(v)
 	require.NoError(t, err)
 
 	currentBinary := uploadVersion(ctx, t, c, c.All(), clusterupgrade.MainVersion)
